requirements.txt
Flask==3.0.3
PyMySQL==1.1.1
SQLAlchemy==2.0.36
google-cloud-vision==3.8.0
google-cloud-storage==2.18.2
numpy==2.0.2
appengine-python-standard==1.1.4
gunicorn